Module Source.Bank

val add_balance : Players.player -> int -> Players.player

add_balance player amount returns a player with amount added to their current balance

val final_balance : Players.player -> int

final_balance player returns the final balance of the player which includes taking into account their current balance, debt, house, and life tiles

val payraise : Players.player -> Players.player

payraise player returns a player with $10,000 added to their current pay_raise

val payday : Players.player -> Players.player

payday player returns a player with their current career's pay and the player's pay_raise (pay + pay_raise is limited at their career's salary max) added to their current balance

val tax : Players.player -> Players.player

tax player returns a player with their career's taxes subtracted from their current balance

val pay_college : Players.player -> Players.player

pay_college player returns a player with 100,000 added to their current debt

val calculate_loans : Players.player -> Players.player

calculate_loans player returns a player with continous loans (loans are 20,000 added to their current balance and 25,000 added to their current debt) added until their account balance is positive