Module Source.Gamestate

type gamestate = {

current_player : Players.player;

players : Players.player list;

tiles : Tiles.tiles list;

deck : Cards.cards list;

graphics : bool;

}

val turn : gamestate -> unit

turn st prints the winner of the Game to terminal

val player_winner : Players.player list -> Players.player -> Players.player

player_winner lst returns the winner of the game. Requires the game is over

val gameover : Players.player list -> bool

gamover lst returns true if the game is over else false

val normalize_text : string -> string

normalize_text s returns s with whitespace trimmed and all lowercase

val print_iter : (int -> unit) -> int -> int -> bool -> unit

print_iter pfun acc cap graphic prints pfun if graphic is true else unit is returned

val get_tile : int -> Tiles.tiles list -> Tiles.tiles

get_tile returns the tile in tiles at given index index. Raises : Failure if list is too short and Invalid Argument if n is negative.

val next_player : Players.player -> Players.player list -> Players.player

next_player current_player players returns the player whose turn is after current_player

val finished : Players.player -> bool

finished player returns true if the player has reached the end of the board and has retired and returns false if player is still playing.

val change_index_board : Players.player -> Players.player * int

change_index_board p returns a pair of the player with an updated location on the board and the number that they spun

val final_tile_index : int

final_tile_index is the index of the last tile on the board