Module Source.Printing

val print_payday : int -> unit

print_payday i prints the row of the payday graphic corresponding to row i

val print_wedding : int -> unit

print_wedding i prints the row of the wedding graphic corresponding to row i

val print_life : int -> unit

print_life i prints the row of the life graphic corresponding to row i

val print_house : int -> unit

print_house i prints the row of the house graphic corresponding to row i

val print_lawsuit : int -> unit

print_lawsuit i prints the row of the lawsuit graphic corresponding to row i